financial_research_agent/requirements.txt
google-generativeai>=0.7.0
langgraph>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
langchain-community>=0.2.0
yfinance>=0.2.40
requests>=2.31.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.20.0
beautifulsoup4>=4.12.0
